Update Transaction Class and Date Class
Created the header and cpp for the transaction list. I modified the existing member list class to work for the transaction list class because I assumed that the member list had all the features you would want me to include. Renamed FindNode to FindMemberNode and FindItemNode. I also added the FindItemNode function and overloaded it to search using either and item name or an item number.